day1_math_agent/requirements.txt
langchain==0.2.0
langchain-openai==0.1.7
python-dotenv==1.0.1
